refactor(bench): complete the adapter split — metrics.py becomes stream_envelope.py
#2163 lifted the metrics cluster out of `stella_harbor/__init__.py` and
stopped there, at 1997 lines. #2153 asked for the whole seam, and it was
right to: the journal-parsing half was left behind in the god file, split
from the metrics that consume it.

`_json_dicts_from_line` and `_stream_to_envelope` join their own callees.
The module is renamed `stream_envelope.py`, because that is now what it is:
the durable journal → run envelope → flat trial metrics path, the reading
half of the pipeline whose writing half `atif.py` already owns
(envelope → ATIF trajectory). The two sit side by side instead of one being
tangled into the adapter.

The seam is exact, which is why it is worth taking now rather than later:
`_stream_to_envelope`'s only dependency outside the moved set is
`fold_witness_observations`, and `_valid_nonnegative_number`'s only two
callers were inside it — so both leave `__init__.py` entirely rather than
becoming re-exports. What the adapter still calls
(`_extract_metrics`, `_load_json_object`, `_stream_to_envelope`) is
re-exported; `_sum_step_usage` is reached only by the tests.

`__init__.py` drops 1997 → 1834, the figure #2153 predicted.

Moved verbatim; no behavior change. The renamed module keeps its
`_FIXED_ADAPTER_SOURCE_PATHS` entry, so the launcher's fail-closed tree
comparison stays exact.

Pure refactor: no witness test, and no test was deleted.

Closes #2153
